Refactor bars and overview
