+++ frontend/views/components/Graphs/Bars.vue
@@ -1,48 +1,223 @@
 <template lang='pug'>
-.has-text-centered
-  p.c-description
-    strong.is-size-5.is-block.c-nowrap {{currency}} {{mincome}}
-    i18n min income
+.c-chart-wrapper
+  svg(
+    width='100%'
+    :viewBox='`0 0 ${ratioX} ${ratioY}`'
+    preserveAspectRatio='xMidYMid meet'
+    aria-labelledby='title'
+    role='img'
+    ref='graph'
+  )
+    // Move graph origin to the middle
+    g(:transform='`translate(0,${middle})`')
+      // Animate using scale from the middle
+      g.g-animate(:style='`transform: scale3d(1,${ ready ? 1 : 0 },1)`')
+        g.graph-bar(
+          v-for='(member, index) in members'
+          :transform='positionX(index)'
+        )
+          // Total needed or total pledge bars
+          path(
+            v-for='(values, index) in [member.total, member.amount]'
+            :class='color(values, index)'
+            :d='roundedRect(values >= 0, 0, positionY(values), width, height(values), width > 15 ? 3 : 1)'
+          )
 
-  div(v-if='history.length')
-    .c-graph
-      span.c-graph-item(
-        v-for='(percentage, index) in history'
-        :key='`percentage-${index}`'
-        :style='{\
-          transform: `scaleY(${getHeight(percentage)})`,\
-          opacity: getOpacity(percentage)\
-        }'
+      // Base with $0 on top of bars
+      line(x1='0' y1='0' :x2='ratioX' y2='0' stroke='#dbdbdb' stroke-width='1')
+
+      // Surplus line on top of bars
+      g(
+        v-for='scaleLine in createScale'
+        :transform='`translate(0,${-scaleLine.position})`'
       )
+        line.g-animate.g-animate-delay(
+          :style='`transform: scale3d(${ready ? 1 : 0},1,1)`'
+          x1='0'
+          y1='0'
+          :x2='ratioX'
+          y2='0'
+          stroke='#dbdbdb'
+          stroke-width='1'
+          stroke-dasharray='1'
+        )
 
-    p.c-description
-      strong.is-block.c-nowrap
-        | + {{getTotal()}}
-        i18n distributed
-      i18n on the last
-      |  {{history.length}}
-      i18n months
-
-  p(v-else='') We are on the first month.
+  .tag.mincome(:style='middleTag') {{ base }}
+  .tag.g-animate-opacity.g-animate-delay(
+    v-for='scaleLine in createScale'
+    :style='{ opacity: ready ? 1 : 0, transform: "translate(0," + (-scaleLine.position + middle - labelPadding) + "px)" }'
+  ) {{ scaleLine.label }}
 </template>
 
 <script>
+import { mapGetters } from 'vuex'
+import { debounce } from '@utils/giLodash.js'
+import { TABLET } from '@view-utils/breakpoints.js'
+import currencies from '@view-utils/currencies.js'
+
 export default {
   name: 'Bars',
   props: {
-    history: Array,
-    currency: String,
-    mincome: Number
+    totals: Array,
+    members: Array
+  },
+  data: () => ({
+    ratioWidthPadding: 1.5,
+    ratioX: 526,
+    ratioY: 160,
+    labelPadding: 10,
+    labelwidth: 40,
+    availableWidth: 474,
+    maxWidth: 48,
+    ready: false,
+    isMobile: false
+  }),
+  mounted () {
+    window.addEventListener('resize', this.handleResize)
+    this.handleResize()
+    setTimeout(() => { this.ready = true }, 0)
+  },
+  beforeDestroy: function () {
+    window.removeEventListener('resize', this.handleResize)
+  },
+  computed: {
+    ...mapGetters([
+      'groupProfiles',
+      'groupSettings'
+    ]),
+    base () {
+      return this.currency(0)
+    },
+    currency () {
+      return currencies[this.groupSettings.mincomeCurrency].displayWithCurrency
+    },
+    membersNumber () {
+      return this.members.length
+    },
+    // Graphic proportions
+    width () {
+      return Math.min(this.availableWidth / this.membersNumber / this.ratioWidthPadding, this.maxWidth)
+    },
+    max () {
+      return Math.max.apply(Math, this.totals)
+    },
+    min () {
+      return Math.min.apply(Math, this.totals)
+    },
+    middle () {
+      return this.calculRatioY(this.max)
+    },
+    middleTag () {
+      return { transform: 'translate(0,' + (this.middle - this.labelPadding) + 'px)' }
+    },
+    createScale () {
+      const range = this.max - this.min
+      const maxScalesCount = 4
+      const roundedTickRange = this.toPrecision(range / maxScalesCount)
+      const scales = []
+      let scale = Math.ceil(this.min / roundedTickRange) * roundedTickRange
+      while (scale <= this.max) {
+        let label = this.currency(Math.abs(scale))
+        if (scale < 0) label = '-' + label
+        // Add scale label and positition
+        scales.push({ label: label, position: this.calculRatioY(scale) })
+        scale = scale + roundedTickRange
+      }
+      return scales
+    }
   },
   methods: {
-    getOpacity (percentage) {
-      return percentage > 1 ? '' : percentage
+    handleResize: debounce(function () {
+      if (this.$refs.graph) {
+        this.isMobile = this.verifyIsMobile()
+        this.ratioX = this.$refs.graph.clientWidth
+        this.availableWidth = this.ratioX - this.labelwidth
+      }
+    }, 100),
+
+    toPrecision (nbr) {
+      if (typeof nbr !== 'number') return 0
+      if (nbr === 0) return 0
+      // log only for positive number
+      const num = Math.abs(nbr)
+      // Remove floating number for large number to calcul the precision
+      if (nbr > 10) nbr = parseInt(nbr)
+      // Increase precision for small value
+      const precision = nbr.toString().length > 3 ? 2 : 1
+      // Rounding technic
+      const digits = Math.ceil(Math.log(num) / Math.LN10)
+      const factor = Math.pow(10, precision - digits)
+      let result = Math.round(num * factor, 0) / factor
+      // Remove floating number for large number
+      if (num > 1000) result = parseInt(result)
+      // Bring sign back
+      return nbr > 0 ? result : -result
     },
-    getHeight (percentage) {
-      return percentage > 1.25 ? 1.25 : percentage
+
+    verifyIsMobile () {
+      return window.innerWidth < TABLET
     },
-    getTotal () {
-      return '$4500'
+
+    calculRatioY (y) {
+      return this.ratioY / (this.max + Math.abs(this.min)) * y
+    },
+
+    height (delta) {
+      return this.calculRatioY(Math.abs(delta))
+    },
+
+    positionX (index) {
+      const marginLeft = Math.min((this.availableWidth - (this.width * this.membersNumber)) / this.membersNumber, 28)
+      const maxWidth = this.membersNumber * (this.width + marginLeft)
+      const positionX = maxWidth / this.membersNumber * index
+      const offset = (this.availableWidth - maxWidth) / 2 - marginLeft
+      return `translate(${positionX + marginLeft + offset}, 0)`
+    },
+
+    positionY (delta) {
+      // If delta is positive, the bar start at 0 otherw
+      return delta >= 0 ? -(this.calculRatioY(delta)) : 0
+    },
+
+    color (delta, surplus) {
+      if (delta >= 0) return surplus ? 'g-positive' : 'g-surplus'
+      else return surplus ? 'g-negative' : 'g-needed'
+    },
+
+    roundedRect (up, x, y, w, h, r) {
+      let retval = 'M' + (x + r) + ',' + y
+      retval += 'h' + (w - 2 * r)
+      // Top right corner
+      if (up) retval += 'a' + r + ',' + r + ' 0 0 1 ' + r + ',' + r // Arc
+      else {
+        // Square corner
+        retval += 'h' + r
+        retval += 'v' + r
+      }
+      retval += 'v' + (h - 2 * r)
+      // Bottom right corner
+      if (up) {
+        // Square corner
+        retval += 'v' + r
+        retval += 'h' + -r
+      } else retval += 'a' + r + ',' + r + ' 0 0 1 ' + -r + ',' + r // Arc
+      retval += 'h' + (2 * r - w)
+      // Bottom left corner
+      if (up) {
+        // Square corner
+        retval += 'h' + -r
+        retval += 'v' + -r
+      } else retval += 'a' + r + ',' + r + ' 0 0 1 ' + -r + ',' + -r // Arc
+      retval += 'v' + (2 * r - h)
+      // Top left corner
+      if (up) retval += 'a' + r + ',' + r + ' 0 0 1 ' + r + ',' + -r // Arc
+      else {
+        // Square corner
+        retval += 'v' + -r
+        retval += 'h' + r
+      }
+      retval += 'z'
+      return retval
     }
   }
 }
@@ -51,45 +226,54 @@ export default {
 <style lang="scss" scoped>
 @import "@assets/style/_variables.scss";
 
-$barHeight: $spacer*2.5;
-$barMarginBottom: $spacer-sm;
-
-.c-description {
-  line-height: 1.2;
-  margin-bottom: $spacer;
-}
-
-.c-nowrap {
-  white-space: nowrap;
-}
-
-.c-graph {
+.c-chart-wrapper {
   display: flex;
+  flex-wrap: wrap;
   position: relative;
-  margin: $spacer 0 $barMarginBottom;
-
-  &::after {
-    content: "";
-    position: absolute;
-    bottom: $barHeight;
-    left: 0;
-    width: 100%;
-    border-top: 1px dashed $background;
-  }
 }
 
-.c-graph-item {
-  width: $spacer;
-  margin: 0 $spacer-xs;
-  height: $barHeight;
-  background: $primary_0;
-  display: block;
-  transform: scaleY(0);
-  transform-origin: 0 100%;
-  animation: scaleUp 750ms forwards;
+.tag {
+  position: absolute;
+  right: 0;
+  font-size: 12px;
+  background-color: #fff;
+  color: var(--text_1);
+  text-align: right;
+  padding-left: 10px;
 }
 
-@keyframes scaleUp {
-  from { transform: scaleY(0); }
+.g-surplus {
+  fill: $success_0;
+}
+
+.g-positive {
+  fill: $primary_0;
+}
+
+.g-negative {
+  fill: $warning_0;
+}
+
+.g-needed {
+  fill: $danger_0;
+}
+
+.g-animate {
+  // TODO: remove important once test are done
+  transition: transform 0.7s ease-out !important;
+}
+
+.g-animate-opacity {
+  // TODO: remove important once test are done
+  transition: opacity 0.7s ease-out !important;
+}
+
+.g-animate-delay {
+  // TODO: remove important once test are done
+  transition-delay: 0.5s !important;
+}
+
+.c-no-activities {
+  display: flex;
 }
 </style>
